pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "sshvault"
version = "0.1.0"
description = "Encrypted local SSH credential manager"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

authors = [
  { name = "ratmist", url = "https://github.com/ratmist" }
]

dependencies = [
  "cryptography>=42.0.0",
  "argon2-cffi>=23.1.0",
  "paramiko>=3.4.0",
  "pyfiglet>=1.0.2"
]

[project.scripts]
sshvault = "sshvault.cli:main"
